+++ features/kasir/components/ui/NotificationBanner.tsx
@@ -1,9 +1,8 @@
 'use client'
 
-import { AlertCircle, X, RotateCcw, CheckCircle } from 'lucide-react'
+import { X } from 'lucide-react'
 import { Button } from '@/components/ui/button'
-
-export type NotificationType = 'error' | 'warning' | 'info' | 'success'
+import { NotificationType, notificationStyles } from '../../lib/constants/uiConfig'
 
 interface NotificationBannerProps {
   /**
@@ -45,55 +44,6 @@ interface NotificationBannerProps {
   ariaLive?: 'polite' | 'assertive'
 }
 
-const notificationStyles: Record<
-  NotificationType,
-  {
-    container: string
-    icon: typeof AlertCircle
-    iconColor: string
-    titleColor: string
-    messageColor: string
-    helpTextColor: string
-    buttonColor: string
-  }
-> = {
-  error: {
-    container: 'bg-red-50 border border-red-200',
-    icon: AlertCircle,
-    iconColor: 'text-red-500',
-    titleColor: 'text-red-900',
-    messageColor: 'text-red-700',
-    helpTextColor: 'text-red-600',
-    buttonColor: 'text-red-500 hover:text-red-700',
-  },
-  warning: {
-    container: 'bg-yellow-50 border border-yellow-200',
-    icon: AlertCircle,
-    iconColor: 'text-yellow-600',
-    titleColor: 'text-yellow-900',
-    messageColor: 'text-yellow-700',
-    helpTextColor: 'text-yellow-600',
-    buttonColor: 'text-yellow-500 hover:text-yellow-700',
-  },
-  info: {
-    container: 'bg-blue-50 border border-blue-200',
-    icon: RotateCcw,
-    iconColor: 'text-blue-500',
-    titleColor: 'text-blue-900',
-    messageColor: 'text-blue-700',
-    helpTextColor: 'text-blue-600',
-    buttonColor: 'text-blue-500 hover:text-blue-700',
-  },
-  success: {
-    container: 'bg-green-50 border border-green-200',
-    icon: CheckCircle,
-    iconColor: 'text-green-500',
-    titleColor: 'text-green-900',
-    messageColor: 'text-green-700',
-    helpTextColor: 'text-green-600',
-    buttonColor: 'text-green-500 hover:text-green-700',
-  },
-}
 
 /**
  * Reusable notification banner component
