+++ features/kasir/components/ui/status-badge.tsx
@@ -1,37 +1,12 @@
 import { cn } from '@/lib/utils'
 import type { TransactionStatus } from '../../types'
+import { statusConfig } from '../../lib/constants/uiConfig'
 
 interface StatusBadgeProps {
   status: TransactionStatus
   className?: string
   'data-testid'?: string
   'aria-label'?: string
-}
-
-const statusConfig: Record<
-  TransactionStatus,
-  { label: string; className: string; description: string }
-> = {
-  active: {
-    label: 'Aktif',
-    className: 'bg-blue-100 text-blue-800 border-blue-200',
-    description: 'Transaksi sedang berjalan',
-  },
-  selesai: {
-    label: 'Selesai',
-    className: 'bg-green-100 text-green-800 border-green-200',
-    description: 'Transaksi telah selesai',
-  },
-  terlambat: {
-    label: 'Terlambat',
-    className: 'bg-red-100 text-red-800 border-red-200',
-    description: 'Transaksi terlambat dikembalikan',
-  },
-  cancelled: {
-    label: 'Dibatalkan',
-    className: 'bg-gray-100 text-gray-800 border-gray-200',
-    description: 'Transaksi dibatalkan',
-  },
 }
 
 export function StatusBadge({
