+++ features/kasir/lib/constants/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Centralized exports for kasir constants
+ * Single import point for all configuration objects
+ */
+
+// Step validation messages
+export * from './stepValidationMessages'
+
+// UI configuration constants
+export * from './uiConfig'
+
+// Workflow configuration constants
+export * from './workflowConfig'
